+++ packages/va-poc/src/views/ProfileView.vue
@@ -16,24 +16,10 @@ watchEffect(() => {
   locale.value = newLocale.value;
 });
 
-// Import the custom elements
-import LWSSourceAdder from "@/components/providers/LWSSourceAdder.ce.vue";
-import LWSSourcesList from "@/components/providers/LWSSourcesList.ce.vue";
 import TranslationTester from "@/components/TranslationTester.ce.vue";
-import LWSProcessList from "@/components/providers/LWSProcessList.ce.vue";
 
-// Register the custom elements
-if (!customElements.get('lws-source-adder')) {
-  customElements.define('lws-source-adder', defineCustomElement(LWSSourceAdder));
-}
-if (!customElements.get('lws-sources-list')) {
-  customElements.define('lws-sources-list', defineCustomElement(LWSSourcesList));
-}
 if (!customElements.get('translation-tester')) {
   customElements.define('translation-tester', defineCustomElement(TranslationTester));
-}
-if (!customElements.get('lws-process-list')) {
-  customElements.define('lws-process-list', defineCustomElement(LWSProcessList));
 }
 </script>
 
@@ -42,10 +28,8 @@ if (!customElements.get('lws-process-list')) {
     <h1>{{ t("welcome") }}</h1>
     <p>{{ t("message") }}</p>
     <BContainer>
-      <lws-source-adder />
-      <lws-sources-list />
-      <lws-process-list />
-      <!-- <translation-tester name="Karel">Nothing to report</translation-tester> -->
+      <translation-tester name="Karel">Nothing to report</translation-tester>
+      Todo: allow editing orgs and sites!
     </BContainer>
   </section>
 </template>
